+++ server/component/BaseModel.php
@@ -249,5 +249,21 @@ abstract class BaseModel
     {
         return $this->db->get_languages();
     }
+
+    /**
+     * get user groups from the database.
+     *
+     *  @retval array
+     *  value int,
+     *  text string
+     */
+    public function get_groups()
+    {
+        $groups = array();
+        foreach ($this->db->select_table("groups") as $group) {
+            array_push($groups, array("value" => intval($group['id']), "text" => $group['name']));
+        }
+        return $groups;
+    }
 }
 ?>